+++ tests/fetching/test_abstract_fetcher.py
@@ -1,4 +1,5 @@
 import pytest
+from rics.mapping import Mapper
 
 from id_translation.fetching import AbstractFetcher, MemoryFetcher, exceptions
 from id_translation.fetching.types import IdsToFetch
@@ -30,3 +31,24 @@ def test_fetch_all_forbidden(data):
 def test_unknown_placeholders(fetcher):
     with pytest.raises(exceptions.UnknownPlaceholderError, match="{'number_of_legs'} not recognized"):
         fetcher.fetch([IdsToFetch("humans", [])], ("id", "number_of_legs"), {"number_of_legs"})
+
+
+@pytest.mark.parametrize(
+    "selective_fetch_all, required, expected",
+    [
+        (True, ("id", "name"), {"animals", "humans"}),
+        (False, (), {"animals", "humans", "big_table", "huge_table"}),
+    ],
+)
+def test_selective_fetch_all(data, selective_fetch_all, required, expected):
+    fetcher: AbstractFetcher[str, int] = MemoryFetcher(
+        data,
+        selective_fetch_all=selective_fetch_all,
+        mapper=Mapper(unmapped_values_action="raise"),
+    )
+    assert set(fetcher.fetch_all(required=required)) == expected
+
+
+def test_test_crashes_with_selective_fetch_all_disabled(data):
+    with pytest.raises(exceptions.UnknownPlaceholderError):
+        MemoryFetcher(data, selective_fetch_all=False).fetch_all(required=("id", "name"))