+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Finished `Translator.load_persistent_instance()` implementation (no longer experimental).
 - The `SqlFetcher.finalize_statement()` method, used to customize fetching behavior programmatically.
 - New INFO-level begin/end log messages for `Translator.translate()`.
-- Raise `ConcurrentOperationError` in `AbstractFetcher` to prevent race conditions.
+- Raise `ConcurrentOperationError` in `AbstractFetchers.fetch()` to prevent race conditions.
+- Limit `AbstractFetcher.fetch_all()` to sources that contain the required placeholders (after mapping) by default.
 
 ### Changed
 - Improve error reporting for unmapped required placeholders; warn about potential override issues.